Reorganize package listing data fetch
Move the fetchin logic to its own component
Simplify the logic to only use string comparison

refs TS-712
