+++ apps/sertra/pages/server/[id].tsx
@@ -7,8 +7,9 @@ import {
   ServerListingData,
   ListingMod,
 } from "../../api/models";
-import { ApiURLs, TsApiURLs } from "../../api/urls";
+import { ApiURLs } from "../../api/urls";
 import { ModCard } from "../../components/ModCard";
+import { FetchListingData } from "../../components/PackageDataFetcher";
 import { ServerInfo } from "../../components/ServerInfo";
 import { ServerInstructions } from "../../components/ServerInstructions";
 import styles from "../../styles/ServerDetail.module.css";
@@ -24,7 +25,7 @@ export const getStaticPaths: GetStaticPaths = async () => {
   return { paths, fallback: "blocking" };
 };
 
-type ServerListingStaticProps = { detail_listing?: ServerListingDetailData };
+type ServerListingStaticProps = { listing_data?: ServerListingDetailData };
 type ServerListingQueryProps = { id: string };
 
 export const getStaticProps: GetStaticProps<
@@ -34,59 +35,20 @@ export const getStaticProps: GetStaticProps<
   // Params is never undefined thanks to NextJS guarantees as long as the file
   // is named appropriately.
   const listingId = context.params!.id; // eslint-disable-line @typescript-eslint/no-non-null-assertion
-  const res = await fetch(ApiURLs.ServerDetail(listingId));
-  const mods_res = await fetch(TsApiURLs.V1Packages("v-rising"));
-  const data = await res.json();
-  const mods_data = await mods_res.json();
-
-  const updated_listing_mods_data = [];
-  for (const mod_ref of data.mods) {
-    const parsed_mod_ref = [];
-    // Split, reverse, put together, split by dash,
-    // take each from dash split and flip it around, put into new array
-    for (const x of mod_ref.split("").reverse().join("").split("-")) {
-      parsed_mod_ref.push(x.split("").reverse().join(""));
-    }
-    const new_mod_data = {
-      // We are accepting null here, just in case the data we have about mods is corrupt or bad
-      name: parsed_mod_ref[1] ?? null,
-      owner: parsed_mod_ref.slice(2).join("-") ?? null,
-      version: parsed_mod_ref[0] ?? null,
-      icon_url: null,
-      description: null,
-    };
-    // We could use the same parser output to match against APIs results
-    // But I'm waiting for opinions
-    for (const mod of mods_data) {
-      if (mod_ref.startsWith(mod.full_name)) {
-        new_mod_data.name = mod.name;
-        new_mod_data.owner = mod.owner;
-        new_mod_data.version = null;
-        for (const version of mod.versions) {
-          if (mod_ref.endsWith(version.version_number)) {
-            new_mod_data.version = version.version_number;
-            new_mod_data.icon_url = version.icon ?? null;
-            new_mod_data.description = version.description ?? null;
-            break;
-          }
-        }
-      }
-    }
-    updated_listing_mods_data.push(new_mod_data);
-  }
+  const fetched = FetchListingData(listingId);
   return {
     props: {
-      detail_listing: data,
-      updated_mod_datas: updated_listing_mods_data,
+      listing_data: (await fetched).listing_data,
+      mods_data: (await fetched).mods_data,
     },
     revalidate: 10,
   };
 };
 
 const ServerDetail: React.FC<{
-  detail_listing: ServerListingDetailData;
-  updated_mod_datas: ListingMod[];
-}> = ({ detail_listing, updated_mod_datas }) => {
+  listing_data: ServerListingDetailData;
+  mods_data: ListingMod[];
+}> = ({ listing_data, mods_data }) => {
   return (
     <div className={styles.container}>
       <div className={styles.breadcrumb}>
@@ -96,7 +58,7 @@ const ServerDetail: React.FC<{
       </div>
 
       <div className={styles.headerRow}>
-        <h1 className={styles.listingTitle}>{detail_listing.name}</h1>
+        <h1 className={styles.listingTitle}>{listing_data.name}</h1>
         <button className={styles.joinServerButton}>
           <Logo />
           Join Server
@@ -107,13 +69,13 @@ const ServerDetail: React.FC<{
         <div className={styles.columnLeft}>
           <section>
             <h2 className={styles.sectionTitle}>Description</h2>
-            <p className={styles.description}>{detail_listing.description}</p>
+            <p className={styles.description}>{listing_data.description}</p>
           </section>
 
           <section>
             <h2 className={styles.sectionTitle}>Mods</h2>
             <div>
-              {updated_mod_datas.map((x) => (
+              {mods_data.map((x) => (
                 <ModCard key={x.name} {...x} />
               ))}
             </div>
@@ -121,7 +83,7 @@ const ServerDetail: React.FC<{
         </div>
 
         <div className={styles.columnRight}>
-          <ServerInfo {...detail_listing} />
+          <ServerInfo {...listing_data} />
           <ServerInstructions />
         </div>
       </div>
